Remove specific macOS configuration file
