--- Sources/Mac/Extensions/Configuration+Imaginary.swift
@@ -1,34 +1,0 @@
-import Cocoa
-import Cache
-
-public extension Configuration {
-  public static var preConfigure: ((NSImageView) -> Void)? = { imageView in
-    imageView.layer?.opacity = 0.0
-  }
-
-  public static var transitionClosure: ((NSImageView, NSImage) -> Void) = { imageView, newImage in
-    guard let oldImage = imageView.image, imageView.window?.inLiveResize == false else {
-      imageView.image = newImage
-      return
-    }
-
-    let animation = CABasicAnimation(keyPath: "contents")
-    animation.duration = 0.25
-    animation.fromValue = oldImage.cgImage
-    animation.toValue = newImage.cgImage
-    imageView.wantsLayer = true
-    imageView.layer?.add(animation, forKey: "transitionAnimation")
-    imageView.image = newImage
-  }
-
-  public static var postConfigure: ((_ imageView: NSImageView) -> Void)? = { imageView in
-    let animation = CABasicAnimation(keyPath: "opacity")
-    animation.fromValue = imageView.layer?.opacity
-    animation.toValue = 1.0
-    imageView.wantsLayer = true
-    imageView.layer?.add(animation, forKey: "fadeAnimation")
-    imageView.layer?.opacity = 1.0
-  }
-
-  public static var track: ((URL?, Error?, Int) -> Void)?
-}
