+++ Sources/Shared/Configuration.swift
@@ -30,9 +30,7 @@ public struct Configuration {
 
   /// Pre configure imageView before setting image
   public var preConfigure: ((ImageView) -> Void)? = { imageView in
-    #if os(iOS) || os(tvOS)
-      imageView.layer.opacity = 0.0
-    #endif
+    imageView.layer.opacity = 0.0
   }
 
   /// Animation when new image is set
@@ -49,6 +47,19 @@ public struct Configuration {
       animation.toValue = newImage.cgImage
       imageView.layer.add(animation, forKey: "transitionAnimation")
       imageView.image = newImage
+    #elseif os(OSX)
+      guard let oldImage = imageView.image, imageView.window?.inLiveResize == false else {
+        imageView.image = newImage
+        return
+      }
+
+      let animation = CABasicAnimation(keyPath: "contents")
+      animation.duration = 0.25
+      animation.fromValue = oldImage.cgImage
+      animation.toValue = newImage.cgImage
+      imageView.wantsLayer = true
+      imageView.layer?.add(animation, forKey: "transitionAnimation")
+      imageView.image = newImage
     #endif
   }
 
@@ -60,6 +71,13 @@ public struct Configuration {
       animation.toValue = 1.0
       imageView.layer.add(animation, forKey: "fadeAnimation")
       imageView.layer.opacity = 1.0
+    #elseif os(OSX)
+      let animation = CABasicAnimation(keyPath: "opacity")
+      animation.fromValue = imageView.layer?.opacity
+      animation.toValue = 1.0
+      imageView.wantsLayer = true
+      imageView.layer?.add(animation, forKey: "fadeAnimation")
+      imageView.layer?.opacity = 1.0
     #endif
   }
